--- Excel_Adapter/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ClosedXML" version="0.95.0" targetFramework="net461" />
-  <package id="DocumentFormat.OpenXml" version="2.7.2" targetFramework="net461" />
-  <package id="ExcelNumberFormat" version="1.0.10" targetFramework="net461" />
-  <package id="Microsoft.CSharp" version="4.7.0" targetFramework="net461" />
-  <package id="System.IO.FileSystem.Primitives" version="4.0.1" targetFramework="net461" />
-  <package id="System.IO.Packaging" version="4.0.0" targetFramework="net461" />
-</packages>
