--- Excel_Engine/Compute/ParseEnum.cs
@@ -1,68 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2022, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using BH.Engine.Reflection;
-using BH.oM.Base.Attributes;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BH.Engine.Excel
-{
-    public static partial class Compute
-    {
-        /*******************************************/
-        /**** Public Methods                    ****/
-        /*******************************************/
-
-        public static T ParseEnum<T>(string value)
-        {
-            return (T)ParseEnum(typeof(T), value);
-        }
-
-        /*******************************************/
-
-        public static object ParseEnum(Type enumType, string value)
-        {
-            if (Enum.IsDefined(enumType, value))
-                return Enum.Parse(enumType, value);
-            else
-            {
-                return Enum.GetValues(enumType).OfType<Enum>()
-                    .FirstOrDefault(x => {
-                        FieldInfo fi = enumType.GetField(x.ToString());
-                        DescriptionAttribute[] attributes = fi.GetCustomAttributes(typeof(DescriptionAttribute), false) as DescriptionAttribute[];
-
-                        return attributes != null && attributes.Count() > 0 && attributes.First().Description == value;
-                    });
-            }
-        }
-
-        /*******************************************/
-    }
-}
-
