--- Excel_Engine/Query/ColumnName.cs
@@ -1,61 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2022, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using BH.oM.Base.Attributes;
-using System.ComponentModel;
-
-namespace BH.Engine.Excel
-{
-    public static partial class Query
-    {
-        /*******************************************/
-        /**** Public Methods                    ****/
-        /*******************************************/
-
-        [Description("Returns the letter representing the column under given index, starting from 0 equal to A in Excel.")]
-        [Input("columnIndex", "Index of the column to be returned, starting from 0 equal to A in Excel.")]
-        [Output("column", "Letter representing the column under given index.")]
-        public static string ColumnName(this int columnIndex)
-        {
-            if (columnIndex < 0)
-            {
-                BH.Engine.Base.Compute.RecordError("Negative column index is not allowed.");
-                return null;
-            }
-
-            string columnName = "";
-
-            columnIndex++;
-            while (columnIndex > 0)
-            {
-                int modulo = (columnIndex - 1) % 26;
-                columnName = System.Convert.ToChar('A' + modulo) + columnName;
-                columnIndex = (columnIndex - modulo) / 26;
-            }
-
-            return columnName;
-        }
-
-        /*******************************************/
-    }
-}
-
