+++ Excel_UI.sln
@@ -1,15 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 15 for Windows Desktop
-VisualStudioVersion = 15.0.27130.2024
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32126.317
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Excel_UI", "Excel_UI\Excel_UI.csproj", "{0FAF6FEB-80C3-474A-92BB-C886424CCBC8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Excel_Engine", "Excel_Engine\Excel_Engine.csproj", "{A6884466-297E-4AAA-A232-ABBEC42168D3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Excel_oM", "Excel_oM\Excel_oM.csproj", "{79FAC4B0-77EC-4257-BCBE-CFF37FFF537C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Excel_Adapter", "Excel_Adapter\Excel_Adapter.csproj", "{79D1BABE-DF98-4635-A605-AA396FF58915}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExcelUI_Engine", "ExcelUI_Engine\ExcelUI_Engine.csproj", "{56039ED6-FBD5-4F45-936B-2116399AEEC3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,18 +17,10 @@ Global
 		{0FAF6FEB-80C3-474A-92BB-C886424CCBC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0FAF6FEB-80C3-474A-92BB-C886424CCBC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0FAF6FEB-80C3-474A-92BB-C886424CCBC8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A6884466-297E-4AAA-A232-ABBEC42168D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A6884466-297E-4AAA-A232-ABBEC42168D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A6884466-297E-4AAA-A232-ABBEC42168D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A6884466-297E-4AAA-A232-ABBEC42168D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{79FAC4B0-77EC-4257-BCBE-CFF37FFF537C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79FAC4B0-77EC-4257-BCBE-CFF37FFF537C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79FAC4B0-77EC-4257-BCBE-CFF37FFF537C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79FAC4B0-77EC-4257-BCBE-CFF37FFF537C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{79D1BABE-DF98-4635-A605-AA396FF58915}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79D1BABE-DF98-4635-A605-AA396FF58915}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79D1BABE-DF98-4635-A605-AA396FF58915}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79D1BABE-DF98-4635-A605-AA396FF58915}.Release|Any CPU.Build.0 = Release|Any CPU
+		{56039ED6-FBD5-4F45-936B-2116399AEEC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{56039ED6-FBD5-4F45-936B-2116399AEEC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{56039ED6-FBD5-4F45-936B-2116399AEEC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{56039ED6-FBD5-4F45-936B-2116399AEEC3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
