--- Excel_oM/Requests/CellValuesRequest.cs
@@ -1,46 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2022, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using BH.oM.Data.Requests;
-using System.Collections.Generic;
-using System.ComponentModel;
-
-namespace BH.oM.Adapters.Excel
-{
-    [Description("IRequest that pulls values only from the Excel file.")]
-    public class CellValuesRequest : IRequest
-    {
-        /***************************************************/
-        /****                Properties                 ****/
-        /***************************************************/
-
-        [Description("Name of the worksheet to read from.")]
-        public virtual string Worksheet { get; set; } = "";
-
-        [Description("Cell range to read. If left empty, entire extent of the spreadsheet starting from A1 is used.")]
-        public virtual CellRange Range { get; set; } = null;
-
-        /***************************************************/
-    }
-}
-
-
