--- Excel_oM/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ClosedXML" version="0.95.3" targetFramework="net452" />
-  <package id="DocumentFormat.OpenXml" version="2.7.2" targetFramework="net452" />
-  <package id="ExcelNumberFormat" version="1.0.10" targetFramework="net452" />
-</packages>
